// src/app/shared/enums/uf.enum.ts
export enum Uf {
  RR = 'Roraima',
  AP = 'Amapá',
  AM = 'Amazonas',
  PA = 'Pará',
  AC = 'Acre',
  RO = 'Rondônia',
  TO = 'Tocantins',
  MA = 'Maranhão',
  PI = 'Piauí',
  CE = 'Ceará',
  RN = 'Rio Grande do Norte',
  PB = 'Paraíba',
  PE = 'Pernambuco',
  AL = 'Alagoas',
  SE = 'Sergipe',
  BA = 'Bahia',
  MT = 'Mato Grosso',
  DF = 'Distrito Federal',
  GO = 'Goiás',
  MS = 'Mato Grosso do Sul',
  MG = 'Minas Gerais',
  ES = 'Espirito Santo',
  RJ = 'Rio de Janeiro',
  SP = 'São Paulo',
  PR = 'Paraná',
  SC = 'Santa Catarina',
  RS = 'Rio Grande do Sul',
}

export interface UfOption {
  nome: string;
  valor: string;
}